+++ src/libraries/dis_plugin/DisExtension.cpp
@@ -82,7 +82,7 @@ void DisExtension::AddPlatform( const tic::Platform_ABC& platform )
     case tic::Platform_ABC::broken:     pdu.SetAppearance( 1, false, true ); break;
     case tic::Platform_ABC::destroyed:  pdu.SetAppearance( 3, true, true ); break;
     }
-    const geometry::Point2d position( platform.GetPosition().X(), platform.GetPosition().Y() );
+    const geometry::Point2f position( platform.GetPosition().X(), platform.GetPosition().Y() );
     const geometry::Point2d geocoord = converter_.ConvertToGeo( position );
     pdu.SetPosition( geocoord.Y(), geocoord.X(), platform.GetAltitude(), platform.GetSpeed(), platform.GetHeading() );
